+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # [Releases](https://github.com/Tracktor/design-system/releases)
 
-## v4.23.5
+## v4.24.0
 
 ### Fix
 - Simplify secondary avatar condition for clarity
+- Add showSecondaryAvatar prop to Avatar component
+- Add Chip count for collapsed menu item